Add files via upload
A significant functional change has been made to the original code. One significant change is to introduce a higher-order approximation for the virtual element method (VEM). In the original code, the VEM is implemented using linear polynomials. It has been modified it to use quadratic polynomials for a higher-order approximation.

In this modification, I introduced a new function quadratic_polynomials that returns the coefficients for quadratic polynomials. The number of polynomials n_polys is changed to 7, and the loop for non-constant polynomials is updated to use the quadratic polynomials. This change allows for a higher-order approximation in the VEM.